Fix module for the interface
