+++ pyproject.toml
@@ -9,10 +9,10 @@ readme = "README.md"
 python = "^3.12"
 
 [tool.poetry.scripts]
-pcmkdir = 'path_chronicle.fso_expansion:create_dir_entry'
-pctouch = 'path_chronicle.fso_expansion:create_file_entry'
-pcpathslist = 'path_chronicle.fso_expansion:list_paths_entry'
-pcrmpath = 'path_chronicle.fso_expansion:remove_path_entry'
+pcmkdir = 'path_chronicle.interface:create_dir_entry'
+pctouch = 'path_chronicle.interface:create_file_entry'
+pcpathslist = 'path_chronicle.interface:list_paths_entry'
+pcrmpath = 'path_chronicle.interface:remove_path_entry'
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.2.2"
